requirements.txt
# Core FastAPI dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# LLM clients (optional)
openai>=1.0.0
anthropic>=0.7.0

# Vector search (optional)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
